feat: add order actions and reducer
